Activate service account via gcloud command during publish
Triggers nightly failed during ko step of publish task with the following error:
ERROR: gcloud crashed (AttributeError): 'Credentials' object has no attribute 'private_key'

This error was also recently seen in the Pipelines repo and the fix was
to activate the service account prior to issuing gcloud commands. See
https://github.com/tektoncd/pipeline/pull/2847 for that PR.

This commit introduces the same fix to triggers as was done for pipelines.